packages/mlopspython-inference/pyproject.toml
[tool.poetry]
name = "mlopspython-inference"
version = "0.0.0"
description = ""
authors = ["Guillaume Chervet <guillaume.chervet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
opencv-contrib-python-headless = "4.8.0.76"
keras = "2.12.0"
tensorflow = "2.12.0"
tensorflow-io-gcs-filesystem = "0.27.0"
pillow = "10.0.1"

[tool.poetry.group.dev.dependencies]
coverage = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"